continuando a resolver cods em php
resolução de problemas em PHP

+++ Introdução ao Desenvolvimento de Sistemas/lista-de-exercicios-site/php/exercicio1.php
@@ -1,5 +1,5 @@
 <?php
-$valorX = (int) $_GET['valorX'];
+$valorX = (float) $_GET['valorX'];
 
 $valorY = (3 * $valorX) + 2;
